[IC-1] Nitin- Fixed few bugs.

+++ src/main/java/com/tools/service/ImageService.java
@@ -5,6 +5,6 @@ import org.springframework.web.multipart.MultipartFile;
 import java.io.IOException;
 
 public interface ImageService {
-  public String compressImage(MultipartFile multipartFile, int CompressionQuality)
+  public String compressImage(MultipartFile multipartFile, float CompressionQuality)
       throws IOException;
 }
